backend/requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
openai
pydantic<2